+++ R/bayesfactor.R
@@ -1,14 +1,16 @@
 #' Bayes Factors (BF)
 #'
 #' This function compte the Bayes factors (BFs) that are appropriate to the input.
-#' For vectors or single models, it will compute \code{\link[=bayesfactor_savagedickey]{Savage-Dickey's BFs}}.
+#' For vectors or single models, it will compute \code{\link[=bayesfactor_parameters]{BFs for single parameters}},
+#' or is \code{hypothesis} is specified, \code{\link[=bayesfactor_restricted]{BFs for restricted models}}.
 #' For multiple models, it will return the BF corresponding to \code{\link[=bayesfactor_models]{comparison between models}}
 #' and if a model comparison is passed, it will compute the \code{\link[=bayesfactor_inclusion]{inclusion BF}}.
 #' \cr\cr
 #' For a complete overview of these functions, read the \href{https://easystats.github.io/bayestestR/articles/bayes_factors.html}{Bayes factor vignette}.
 #'
 #' @param ... A numeric vector, model object(s), or the output from \code{bayesfactor_models}.
-#' @inheritParams bayesfactor_savagedickey
+#' @inheritParams bayesfactor_parameters
+#' @inheritParams bayesfactor_restricted
 #' @inheritParams bayesfactor_models
 #' @inheritParams bayesfactor_inclusion
 #'
@@ -44,14 +46,15 @@
 #' @export
 bayesfactor <-
   function(...,
-             prior = NULL,
-             direction = "two-sided",
-             hypothesis = 0,
-             effects = c("fixed", "random", "all"),
-             verbose = TRUE,
-             denominator = 1,
-             match_models = FALSE,
-             prior_odds = NULL) {
+           prior = NULL,
+           direction = "two-sided",
+           null = 0,
+           hypothesis = NULL,
+           effects = c("fixed", "random", "all"),
+           verbose = TRUE,
+           denominator = 1,
+           match_models = FALSE,
+           prior_odds = NULL) {
     mods <- list(...)
 
     if (length(mods) > 1) {
@@ -64,10 +67,18 @@ bayesfactor <-
       } else {
         bayesfactor_models(...)
       }
+    } else if (!is.null(hypothesis)) {
+      bayesfactor_restricted(...,
+                             prior = prior,
+                             verbose = verbose,
+                             effects = effects)
     } else {
-      bayesfactor_savagedickey(...,
-        prior = prior, direction = direction,
-        hypothesis = hypothesis, effects = effects,
+      bayesfactor_parameters(
+        ...,
+        prior = prior,
+        direction = direction,
+        null = null,
+        effects = effects,
         verbose = verbose
       )
     }