+++ tests/testthat/test-emmGrid.R
@@ -72,11 +72,11 @@ test_that("emmGrid rope", {
   testthat::expect_equal(xrope$ROPE_Percentage, c(5.53, 0, 1.83), tolerance = 0.1)
 })
 
-test_that("emmGrid bayesfactor_savagedickey", {
+test_that("emmGrid bayesfactor_parameters", {
   testthat::skip_on_travis()
   testthat::skip_on_cran()
   set.seed(4)
-  xsdbf <- bayesfactor_savagedickey(all_, prior = model)
+  xsdbf <- bayesfactor_parameters(all_, prior = model)
   testthat::expect_equal(log(xsdbf$BF), c(-2.5764463544813, 2.00205724074489, -0.235346262395184), tolerance = 1e-4)
   testthat::expect_warning(bayesfactor_savagedickey(all_))
 })